refactor: 레거시 컨트롤러를 어노테이션 기반으로 변경
