--- slipp/src/main/java/slipp/annotationcontroller/AnnotationLoginController.java
@@ -1,34 +1,0 @@
-package slipp.annotationcontroller;
-
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.controller.UserSessionUtils;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-@Controller
-public class AnnotationLoginController {
-    @RequestMapping(value = "/users/login", method = RequestMethod.POST)
-    public String logIn(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        String userId = req.getParameter("userId");
-        String password = req.getParameter("password");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            req.setAttribute("loginFailed", true);
-            return "/user/login.jsp";
-        }
-        if (user.matchPassword(password)) {
-            HttpSession session = req.getSession();
-            session.setAttribute(UserSessionUtils.USER_SESSION_KEY, user);
-            return "redirect:/";
-        } else {
-            req.setAttribute("loginFailed", true);
-            return "/user/login.jsp";
-        }
-    }
-}
