--- slipp/src/main/java/slipp/annotationcontroller/AnnotationProfileController.java
@@ -1,24 +1,0 @@
-package slipp.annotationcontroller;
-
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class AnnotationProfileController {
-    @RequestMapping(value = "/users/profile", method = RequestMethod.GET)
-    public String showProfile(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
-        }
-        req.setAttribute("user", user);
-        return "/user/profile.jsp";
-    }
-}
