+++ slipp/src/main/java/slipp/controller/BaseController.java
@@ -1,0 +1,42 @@
+package slipp.controller;
+
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.JspView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
+import slipp.domain.User;
+import slipp.support.db.DataBase;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@Controller
+public class BaseController {
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public String index(HttpServletRequest req, HttpServletResponse resp) {
+        req.setAttribute("users", DataBase.findAll());
+        return "home.jsp";
+    }
+
+    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
+    public ModelAndView signupForm(HttpServletRequest req, HttpServletResponse resp) {
+        return new ModelAndView(new JspView("/user/form.jsp"));
+    }
+
+    @RequestMapping(value = "/users/loginForm", method = RequestMethod.GET)
+    public ModelAndView loginForm(HttpServletRequest req, HttpServletResponse resp) {
+        return new ModelAndView(new JspView("/user/login.jsp"));
+    }
+
+    @RequestMapping(value = "/users/updateForm", method = RequestMethod.GET)
+    public String updateForm(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        }
+        req.setAttribute("user", user);
+        return "/user/updateForm.jsp";
+    }
+}
