--- slipp/src/main/java/slipp/controller/UpdateUserController.java
@@ -1,27 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.asis.Controller;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class UpdateUserController implements Controller {
-    private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
-
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        User user = DataBase.findUserById(req.getParameter("userId"));
-        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
-            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
-        }
-
-        User updateUser = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
-                req.getParameter("email"));
-        log.debug("Update User : {}", updateUser);
-        user.update(updateUser);
-        return "redirect:/";
-    }
-}
